requirements.txt
streamlit
pandas
plotly
groq
streamlit-calendar
psycopg2-binary
SQLAlchemy
python-dotenv
supabase
yt-dlp
